Numba formal integral (#1549)
* Readded the numba formal integral changes

* Update base.py to match main

* Reverted formal integral instantiation to be more similar to the original C version for consistency

* Moved numba implimentation to the montecarlo_numba folder since this is a numba implementation

* Removed extra ravel

* Added back old formal integral for testing

* Added test for function in the numba formal integral

* Changed the way the shell size is computed, should be correct now

* nevermind

* Moved numba formal integral tests so that paths work out for testing framework

* Moved pointer incrementation to the end of the loop

* Fixed typo

* Switched tests to use numba models

* numba models instantiate with proper data types now

* Fixed numba_formal_integral tests to instantiate numba models with numba compatable types

* Added a couple docstrings with some notes about future improvements

* Streamlined the interface slightly to better catch bugs

* Added some debug statements to watch array sizes in tests

* Changed way size_line and size_shell are created using tau_sobolevs

* Changed the most inner for-loop to be more similar to the C code and updated the njit dict to no longer use fastmath

* Added back in the check for interpolate shells

* Added back in old code for interpolate shells and now computes over the interpolated shells

* Fixed interpolate shells function and made sure dataframes are properly returned when computing the intergral

* Fixed bug in counter for Jred_lu pointer

* More explicit about typing.  the value for SIGMA_THOMSON now comes from the numba config (which is should have in the past).  Should now match C code within test criteria

* Finished tests for the numba formal integral helper functions

* Added some docstrings and removed some commented out code

* Removed old import of C formal integral

* Removed deprecated formal integral import

* Changed test to generate a new numba model as to not interfere with other tests

* removed old C interface

* Updated tests and environment to exclude compilers and Cython

* Reformatted conftest so black doesn't complain

* Removed commented out lines, made cleaner

* Added back montecarlo tests since it apepars some r_packet tests are at least set up here.  Also, readded the setup_package.py file

* Removed commented out packages from the environment file

* Slight optimization and readability improvements

* Further optimization, now pre-computes indices of next resonance point

* Removed old binary search, made formatting fixed
