--- tardis/montecarlo/src/omp_helper.h
@@ -1,7 +1,0 @@
-#ifdef WITHOPENMP
-  #include <omp.h>
-#else
-int omp_get_num_threads(){
-  return 1;
-}
-#endif
