--- tardis/montecarlo/src/randomkit/randomkit.h
@@ -1,42 +1,0 @@
-/* Random kit 1.6 */
-
-/*
- * Anyone who attempts to generate random numbers by deterministic means is, 
- * of course, living in a state of sin.
- * -- John von Neumann 
- */
-
-/*
- * Copyright (c) 2003-2006, Jean-Sebastien Roy (js@jeannot.org)
- * 
- * Permission is hereby granted, free of charge, to any person obtaining a
- * copy of this software and associated documentation files (the
- * "Software"), to deal in the Software without restriction, including
- * without limitation the rights to use, copy, modify, merge, publish,
- * distribute, sublicense, and/or sell copies of the Software, and to
- * permit persons to whom the Software is furnished to do so, subject to
- * the following conditions:
- * 
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- * 
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS
- * OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
- * MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
- * IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY
- * CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT,
- * TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
- * SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
- */
-
-/* @(#) $Jeannot: randomkit.h,v 1.29 2006/02/19 14:44:14 js Exp $ */
- 
-#ifndef _RANDOMKIT_
-#define _RANDOMKIT_
-
-#include "rk_mt.h"
-#include "rk_sobol.h"
-#include "rk_primitive.h"
-#include "rk_isaac.h"
-
-#endif /* _RANDOMKIT_ */
