+++ tardis_env3.yml
@@ -12,8 +12,6 @@ dependencies:
   - astropy=3
   - numba=0.50
   - numexpr
-  - Cython=0.29
-  - compilers
 
   # Plasma
   - networkx
